Move test-dummy-feature creation code to a more general place
Subsequently this allows reuse in other tests
