+++ ui/org.eclipse.pde.ui.tests/src/org/eclipse/pde/ui/tests/launcher/FeatureBasedLaunchTest.java
@@ -21,34 +21,33 @@ import static java.util.Map.entry;
 import static java.util.Map.ofEntries;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.eclipse.pde.internal.core.ICoreConstants.DEFAULT_VERSION;
+import static org.eclipse.pde.ui.tests.util.ProjectUtils.addIncludedFeature;
+import static org.eclipse.pde.ui.tests.util.ProjectUtils.addIncludedPlugin;
+import static org.eclipse.pde.ui.tests.util.ProjectUtils.addRequiredFeature;
+import static org.eclipse.pde.ui.tests.util.ProjectUtils.addRequiredPlugin;
+import static org.eclipse.pde.ui.tests.util.ProjectUtils.createFeatureProject;
 import static org.eclipse.pde.ui.tests.util.ProjectUtils.createPluginProject;
 import static org.eclipse.pde.ui.tests.util.TargetPlatformUtil.bundle;
 import static org.eclipse.pde.ui.tests.util.TargetPlatformUtil.resolution;
 import static org.osgi.framework.Constants.REQUIRE_BUNDLE;
 import static org.osgi.framework.Constants.RESOLUTION_OPTIONAL;
 
-import java.io.PrintWriter;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.*;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
-import org.eclipse.core.resources.*;
-import org.eclipse.core.runtime.*;
+import org.eclipse.core.runtime.CoreException;
+import org.eclipse.core.runtime.Platform;
 import org.eclipse.debug.core.*;
-import org.eclipse.jface.operation.IRunnableWithProgress;
 import org.eclipse.osgi.service.resolver.BundleDescription;
 import org.eclipse.pde.core.plugin.*;
 import org.eclipse.pde.core.target.NameVersionDescriptor;
-import org.eclipse.pde.internal.core.*;
-import org.eclipse.pde.internal.core.feature.FeatureChild;
-import org.eclipse.pde.internal.core.feature.WorkspaceFeatureModel;
-import org.eclipse.pde.internal.core.ifeature.*;
+import org.eclipse.pde.internal.core.DependencyManager;
+import org.eclipse.pde.internal.core.ifeature.IFeature;
 import org.eclipse.pde.internal.launching.launcher.BundleLauncherHelper;
-import org.eclipse.pde.internal.ui.wizards.feature.AbstractCreateFeatureOperation;
-import org.eclipse.pde.internal.ui.wizards.feature.FeatureData;
 import org.eclipse.pde.launching.EclipseApplicationLaunchConfiguration;
 import org.eclipse.pde.launching.IPDELauncherConstants;
+import org.eclipse.pde.ui.tests.util.ProjectUtils.CoreConsumer;
 import org.eclipse.pde.ui.tests.util.TargetPlatformUtil;
 import org.junit.*;
 import org.junit.rules.TemporaryFolder;
@@ -1386,120 +1385,9 @@ public class FeatureBasedLaunchTest extends AbstractLaunchTest {
 
 	// --- utility methods ---
 
-	private static interface CoreConsumer<E> {
-		void accept(E e) throws CoreException;
-	}
-
-	private static void createFeatureProject(String id, String version, CoreConsumer<IFeature> featureSetup)
-			throws Exception {
-		createFeature(id, version, id + "_" + version.replace('.', '_'), featureSetup);
-	}
-
-	private static IFeature createFeature(String id, String version, String projectName,
-			CoreConsumer<IFeature> featureSetup) throws Exception {
-		FeatureData featureData = new FeatureData();
-		featureData.id = id;
-		featureData.version = version;
-
-		IWorkspaceRoot workspaceRoot = ResourcesPlugin.getWorkspace().getRoot();
-		IProject project = workspaceRoot.getProject(projectName);
-		IPath location = workspaceRoot.getLocation().append(project.getName());
-
-		IRunnableWithProgress operation = new AbstractCreateFeatureOperation(project, location, featureData, null) {
-			@Override
-			protected void configureFeature(IFeature feature, WorkspaceFeatureModel model) throws CoreException {
-				featureSetup.accept(feature);
-			}
-
-			@Override
-			protected void openFeatureEditor(IFile manifestFile) {
-				// don't open in headless tests
-			}
-		};
-		operation.run(new NullProgressMonitor());
-
-		flushPendingResourceChangeEvents(project);
-
-		FeatureModelManager featureModelManager = PDECore.getDefault().getFeatureModelManager();
-		return featureModelManager.getFeatureModel(project).getFeature();
-	}
-
-	private static void flushPendingResourceChangeEvents(IProject project) throws CoreException {
-		// The create-feature operation above creates/modifies the feature.xml,
-		// but not all resource-change-events generated by the operation are
-		// handled when the operation exists. If these change-events are not
-		// handled now they are handled asynchronously by the Autobuild-job. The
-		// problem is that the Feature-model is reset/reload while processing
-		// the changes (which first sets all fields to null/zero and then
-		// re-reads them), but the reload is NOT guarded by corresponding locks.
-		// So if this happens asynchronously and the model is read inbetween the
-		// model state could be inconsistent, which occasionally leads to
-		// test-failure. Furthermore the feature-models are only added to the
-		// FeatureModelManager while processing the resource-changes.
-		// Consequently it has to be ensured that all pending resource change
-		// events are properly processed now and in this thread:
-		// -> Building the project ensures this
-		ResourcesPlugin.getWorkspace().run(m -> project.build(IncrementalProjectBuilder.FULL_BUILD, null), null);
-	}
-
-	private static void addRequiredPlugin(IFeature feature, String id, String version, int matchRule)
-			throws CoreException {
-		addImport(feature, id, version, matchRule, IFeatureImport.PLUGIN);
-	}
-
-	private static void addRequiredFeature(IFeature feature, String id, String version, int matchRule)
-			throws CoreException {
-		addImport(feature, id, version, matchRule, IFeatureImport.FEATURE);
-	}
-
-	private static void addImport(IFeature feature, String id, String version, int matchRule, int type)
-			throws CoreException {
-		IFeatureModelFactory factory = feature.getModel().getFactory();
-		IFeatureImport featureImport = factory.createImport();
-		featureImport.setId(id);
-		featureImport.setVersion(version);
-		featureImport.setMatch(matchRule);
-		featureImport.setType(type);
-
-		feature.addImports(new IFeatureImport[] { featureImport });
-	}
-
-	private static FeatureChild addIncludedFeature(IFeature feature, String id, String version) throws CoreException {
-		FeatureChild featureChild = (FeatureChild) feature.getModel().getFactory().createChild();
-		featureChild.setId(id);
-		featureChild.setVersion(version);
-		featureChild.setOptional(false);
-		feature.addIncludedFeatures(new IFeatureChild[] { featureChild });
-		return featureChild;
-	}
-
-	private static IFeaturePlugin addIncludedPlugin(IFeature feature, String id, String version) throws CoreException {
-		IFeaturePlugin featurePlugin = feature.getModel().getFactory().createPlugin();
-		featurePlugin.setId(id);
-		featurePlugin.setVersion(version);
-		featurePlugin.setUnpack(false);
-		feature.addPlugins(new IFeaturePlugin[] { featurePlugin });
-		return featurePlugin;
-	}
-
 	private NameVersionDescriptor targetFeature(String featureId, String featureVersion,
 			CoreConsumer<IFeature> featureSetup) throws Throwable {
-
-		IFeature feature = createFeature(featureId, featureVersion, "tp-feature-temp-project", featureSetup);
-
-		WorkspaceFeatureModel model = (WorkspaceFeatureModel) feature.getModel();
-		IResource resource = model.getUnderlyingResource();
-
-		Path featureDirectory = tpJarDirectory.resolve(Path.of("features", featureId + "_" + featureVersion));
-		Files.createDirectories(featureDirectory);
-		Path featureFile = featureDirectory.resolve(resource.getProjectRelativePath().toString());
-		try (PrintWriter writer = new PrintWriter(Files.newBufferedWriter(featureFile));) {
-			model.save(writer);
-		}
-		IProject project = resource.getProject();
-		project.delete(IResource.FORCE | IResource.ALWAYS_DELETE_PROJECT_CONTENT, null);
-
-		return new NameVersionDescriptor(featureId, featureVersion, NameVersionDescriptor.TYPE_FEATURE);
+		return TargetPlatformUtil.targetFeature(featureId, featureVersion, featureSetup, tpJarDirectory);
 	}
 
 	private void setTargetPlatform(Map<NameVersionDescriptor, Map<String, String>> bundleDescriptions,
