+++ ui/org.eclipse.pde.ui.tests/src/org/eclipse/pde/ui/tests/util/ProjectUtils.java
@@ -24,10 +24,18 @@ import org.eclipse.core.runtime.*;
 import org.eclipse.jdt.core.IJavaProject;
 import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jdt.launching.environments.IExecutionEnvironment;
+import org.eclipse.jface.operation.IRunnableWithProgress;
 import org.eclipse.pde.core.project.IBundleProjectDescription;
 import org.eclipse.pde.core.project.IBundleProjectService;
 import org.eclipse.pde.core.target.NameVersionDescriptor;
+import org.eclipse.pde.internal.core.FeatureModelManager;
+import org.eclipse.pde.internal.core.PDECore;
+import org.eclipse.pde.internal.core.feature.FeatureChild;
+import org.eclipse.pde.internal.core.feature.WorkspaceFeatureModel;
+import org.eclipse.pde.internal.core.ifeature.*;
 import org.eclipse.pde.internal.ui.wizards.IProjectProvider;
+import org.eclipse.pde.internal.ui.wizards.feature.AbstractCreateFeatureOperation;
+import org.eclipse.pde.internal.ui.wizards.feature.FeatureData;
 import org.eclipse.pde.internal.ui.wizards.plugin.NewProjectCreationOperation;
 import org.eclipse.pde.internal.ui.wizards.plugin.PluginFieldData;
 import org.eclipse.pde.ui.IBundleContentWizard;
@@ -191,6 +199,8 @@ public class ProjectUtils {
 		return project;
 	}
 
+	// --- workspace project deletion ---
+
 	/**
 	 * An (intended) {@link org.junit.ClassRule} that deletes all projects from
 	 * the test-workspace before and after all tests are executed.
@@ -245,4 +255,103 @@ public class ProjectUtils {
 		// back-up, should not change anything if everything is properly used
 		IMPORTED_PROJECTS.retainAll(retainedProjects);
 	}
+
+	// --- Feature project creation and setup ---
+
+	public static interface CoreConsumer<E> {
+		void accept(E e) throws CoreException;
+	}
+
+	public static void createFeatureProject(String id, String version, CoreConsumer<IFeature> featureSetup)
+			throws Exception {
+		createFeature(id, version, id + "_" + version.replace('.', '_'), featureSetup);
+	}
+
+	static IFeature createFeature(String id, String version, String projectName, CoreConsumer<IFeature> featureSetup)
+			throws Exception {
+		FeatureData featureData = new FeatureData();
+		featureData.id = id;
+		featureData.version = version;
+
+		IWorkspaceRoot workspaceRoot = ResourcesPlugin.getWorkspace().getRoot();
+		IProject project = workspaceRoot.getProject(projectName);
+		IPath location = workspaceRoot.getLocation().append(project.getName());
+
+		IRunnableWithProgress operation = new AbstractCreateFeatureOperation(project, location, featureData, null) {
+			@Override
+			protected void configureFeature(IFeature feature, WorkspaceFeatureModel model) throws CoreException {
+				featureSetup.accept(feature);
+			}
+
+			@Override
+			protected void openFeatureEditor(IFile manifestFile) {
+				// don't open in headless tests
+			}
+		};
+		operation.run(new NullProgressMonitor());
+
+		flushPendingResourceChangeEvents(project);
+
+		FeatureModelManager featureModelManager = PDECore.getDefault().getFeatureModelManager();
+		return featureModelManager.getFeatureModel(project).getFeature();
+	}
+
+	private static void flushPendingResourceChangeEvents(IProject project) throws CoreException {
+		// The create-feature operation above creates/modifies the feature.xml,
+		// but not all resource-change-events generated by the operation are
+		// handled when the operation exists. If these change-events are not
+		// handled now they are handled asynchronously by the Autobuild-job. The
+		// problem is that the Feature-model is reset/reload while processing
+		// the changes (which first sets all fields to null/zero and then
+		// re-reads them), but the reload is NOT guarded by corresponding locks.
+		// So if this happens asynchronously and the model is read inbetween the
+		// model state could be inconsistent, which occasionally leads to
+		// test-failure. Furthermore the feature-models are only added to the
+		// FeatureModelManager while processing the resource-changes.
+		// Consequently it has to be ensured that all pending resource change
+		// events are properly processed now and in this thread:
+		// -> Building the project ensures this
+		ResourcesPlugin.getWorkspace().run(m -> project.build(IncrementalProjectBuilder.FULL_BUILD, null), null);
+	}
+
+	public static void addRequiredPlugin(IFeature feature, String id, String version, int matchRule)
+			throws CoreException {
+		addImport(feature, id, version, matchRule, IFeatureImport.PLUGIN);
+	}
+
+	public static void addRequiredFeature(IFeature feature, String id, String version, int matchRule)
+			throws CoreException {
+		addImport(feature, id, version, matchRule, IFeatureImport.FEATURE);
+	}
+
+	private static void addImport(IFeature feature, String id, String version, int matchRule, int type)
+			throws CoreException {
+		IFeatureModelFactory factory = feature.getModel().getFactory();
+		IFeatureImport featureImport = factory.createImport();
+		featureImport.setId(id);
+		featureImport.setVersion(version);
+		featureImport.setMatch(matchRule);
+		featureImport.setType(type);
+
+		feature.addImports(new IFeatureImport[] { featureImport });
+	}
+
+	public static FeatureChild addIncludedFeature(IFeature feature, String id, String version) throws CoreException {
+		FeatureChild featureChild = (FeatureChild) feature.getModel().getFactory().createChild();
+		featureChild.setId(id);
+		featureChild.setVersion(version);
+		featureChild.setOptional(false);
+		feature.addIncludedFeatures(new IFeatureChild[] { featureChild });
+		return featureChild;
+	}
+
+	public static IFeaturePlugin addIncludedPlugin(IFeature feature, String id, String version) throws CoreException {
+		IFeaturePlugin featurePlugin = feature.getModel().getFactory().createPlugin();
+		featurePlugin.setId(id);
+		featurePlugin.setVersion(version);
+		featurePlugin.setUnpack(false);
+		feature.addPlugins(new IFeaturePlugin[] { featurePlugin });
+		return featurePlugin;
+	}
+
 }
